requirements.txt
amqp==5.0.6
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
binaryornot==0.4.4
boto3==1.17.97
botocore==1.20.112
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cookiecutter==1.7.3
crispy-tailwind==0.5.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.2
django-allauth==0.44.0
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-js-asset==1.2.2
django-redis==5.0.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-taggit==1.5.1
django-tailwind==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mutagen==1.45.1
oauthlib==3.1.1
Pillow==8.2.0
poyo==0.5.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.5
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.3.1
stripe==2.58.0
text-unidecode==1.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.4.0
